requirements_backup.txt
absl-py==1.4.0
accelerate==1.7.0
aiohttp==3.11.15
albumentations==2.0.7
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blis==1.3.0
bokeh==3.7.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
cmake==3.31.6
colorcet==3.1.0
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.12
dask==2024.12.1
datasets==2.14.4
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.33.1
dill==0.3.7
distributed==2024.12.1
einops==0.8.1
fastai==2.7.19
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
fsspec==2025.3.2
gitpython==3.1.44
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.31.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.17.1
ipython==7.34.0
ipywidgets==7.7.1
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jupyter_client==6.1.12
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
librosa==0.11.0
lightgbm==4.5.0
llvmlite==0.43.0
markdown==3.8
matplotlib==3.10.0
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
notebook==6.5.7
numba==0.60.0
numexpr==2.10.2
numpy==2.0.2
opencv-python==4.11.0.86
optuna==3.6.1
packaging==24.2
pandas==2.2.2
pillow==11.2.1
pip==24.1.2
plotly==5.24.1
protobuf==5.29.4
psutil==5.9.5
PyYAML==6.0.2
pyarrow==18.1.0
pycocotools==2.0.8
pydantic==2.11.4
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==75.2.0
shap==0.47.2
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow-text==2.18.1
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.7.1
transformers==4.52.2
typing_extensions==4.13.2
umap-learn==0.5.7
urllib3==2.4.0
wandb==0.19.11
xgboost==2.1.4
zipp==3.21.0
